Trait bevy_tweening::Lerp

pub trait Lerp {
    type Scalar;

    // Required method
    fn lerp(&self, other: &Self, scalar: &Self::Scalar) -> Self;
}

Describes a type that can linearly interpolate between two points.

Required Associated Types

type Scalar

The scaling type for linear interpolation.

Required Methods

fn lerp(&self, other: &Self, scalar: &Self::Scalar) -> Self

Given self and another point other, return a point on a line running between the two that is scalar fraction of the distance between the two points.

Object Safety

This trait is not object safe.